The glue every core re-typed lives with the vocabulary that owns it
The trace container's column planning moves beside the schema in
missingno-core (a morepork feature, off by default), so the gb and vcs
bridges keep only what their machines observe, and trace profiles stop
routing through the GB crate. Generic code/data-log bookkeeping and the
instruction byte-iterator likewise leave gb for core. ClockRatio — grown a
phase a save state can name — replaces the four hand-rolled f32 audio
accumulators and the SMS's hand-rolled 3:2 dot counter with exact carried
remainders. The Z80 and PSG chips gain an optional inspect feature carrying
the sidebar builders both Sega boards duplicated, and the button-descriptor
helper five cores re-typed joins the ports vocabulary.

+++ crates/systems/vcs/src/debug/frame.rs
@@ -28,7 +28,7 @@ pub(super) fn frame_interval(standard: TvStandard) -> Duration {
         TvStandard::Ntsc => 262.0,
         TvStandard::Pal | TvStandard::Secam => 312.0,
     };
-    Duration::from_secs_f32(lines * 228.0 / crate::tv_standard::master_clock_hz(standard))
+    Duration::from_secs_f32(lines * 228.0 / crate::tv_standard::master_clock_hz(standard) as f32)
 }
 
 /// The picture window shown from the full field the core emits: skip the
